Move detection of shared tiles to separate file, add more tests
https://github.com/brawer/osmviews/issues/30
