+++ cmd/osmviews-builder/histogram_test.go
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2022 Sascha Brawer <sascha@brawer.ch>
+// SPDX-FileCopyrightText: 2025 Sascha Brawer <sascha@brawer.ch>
 // SPDX-License-Identifier: MIT
 
 package main
@@ -8,12 +8,22 @@ import (
 )
 
 func TestFindSharedTiles(t *testing.T) {
+	// Tiles 1 and 3 share the same data offset.
 	shared := findSharedTiles([]uint32{12, 72, 88, 72, 32, 18})
 	if len(shared) != 1 {
 		t.Fatalf("want len(shared) == 1, got %d", len(shared))
 	}
 
-	samples := shared[72].SampleTiles
+	tile, got := shared[72]
+	if !got {
+		t.Fatalf("expected tile with offset=72 among shared tiles")
+	}
+
+	if tile.UseCount != 2 {
+		t.Errorf("expected shared[72].UseCount=2, got %d", tile.UseCount)
+	}
+
+	samples := tile.SampleTiles
 	found := false
 	for _, tile := range samples {
 		if tile == 1 || tile == 3 {
